session-4/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.5
Django==4.1
django-allauth==0.51.0
django-cache-url==3.4.2
django-debug-toolbar==3.6.0
django-extensions==3.2.0
djangorestframework==3.13.1
environs==9.5.0
idna==3.3
iniconfig==1.1.1
marshmallow==3.17.0
mock==4.0.3
mollie-api-python==2.13.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sentry-sdk==1.9.5
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.11
